ci: Add `xclippy`, basic tests, and dependabot (#2)
* ci: Add `xclippy`, basic tests, and dependabot

* Switch to Buildjet

+++ core/src/alu/divrem/utils.rs
@@ -21,10 +21,7 @@ pub fn get_quotient_and_remainder(b: u32, c: u32, opcode: Opcode) -> (u32, u32) 
             (b as i32).wrapping_rem(c as i32) as u32,
         )
     } else {
-        (
-            (b as u32).wrapping_div(c as u32) as u32,
-            (b as u32).wrapping_rem(c as u32) as u32,
-        )
+        (b.wrapping_div(c), b.wrapping_rem(c))
     }
 }
