+++ core/src/syscall/precompiles/edwards/ed_decompress.rs
@@ -226,7 +226,7 @@ impl<E: EdwardsParameters> Syscall for EdDecompressChip<E> {
         // Re-insert sign bit into last bit of Y for CompressedEdwardsY format
         let mut compressed_edwards_y: [u8; COMPRESSED_POINT_BYTES] = y_bytes;
         compressed_edwards_y[compressed_edwards_y.len() - 1] &= 0b0111_1111;
-        compressed_edwards_y[compressed_edwards_y.len() - 1] |= (sign as u8) << 7;
+        compressed_edwards_y[compressed_edwards_y.len() - 1] |= sign << 7;
 
         // Compute actual decompressed X
         let compressed_y = CompressedEdwardsY(compressed_edwards_y);
